Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I would prefer if we can postpone creating Call until we know if we're going to use MediaTransport or not.
> 
> Then I'm not sure I understand the purpose of the OnTransportChanged callback. Does it indicate that a new transport is to be used, or that a transport is somehow reconfigured?

OnTransportChanged callback already exists:
https://cs.chromium.org/chromium/src/third_party/webrtc/pc/jseptransportcontroller.h?l=57&rcl=e769ed90c3599a131166d5a61625c1b64318fb91

it lacks a very detailed documentation, but it simply informs you that you should use given rtp transport for mid.

It's the same for media transport now. (for given mid, you should use given media transport).

The way I understand is that (and what code does) is that there is a new transport & this new transport can be multiplexed on rtp transport.

So it's only called when the transport is created.

I can move construction of the 'call' object to later, but downstream projects cache peer connection in order to speed up the actual call. The cost of creating a thread is probably miniscule, but without proper experimentation it feels like we may not want to introduce such a change.

Please let's decide what to do here, I am concerned about our ability to land this review (or a reworked one) timely. Can we schedule a time to talk f2f?

Patch-set: 2
Reviewer: Gerrit User 7704 <7704@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0