Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think one source of the problem is that RtpTransportControllerSend is used in multiple ways in Call. For one, it is passed to the constructor of AudioSendStream and VideoSendStream. There it would be an improvement to wire it up via the config structs instead, in a similar way as MediaTransport (and it could be left as null with MediaTransport. Then we also need to move construction of RtpTransportControllerSend. That's an essential step for refactored RtpTransport, which we haven't gotten to yet.
 
Agreed, but that's a separate change.

> Then Call also sets up a few observer links, maybe that could be moved to the code that creates Call? For the CallStatsObserver, it's not clear to me what that is for, but looks like the opposite direction of TargettransferRate.

Agreed, there is more than one. I was going for a 'safe change' that wouldn't cause any breakages :) And eventually, when you implement RTP media transport, this issues will have to be addressed. I was basically hoping to defer that.

> If you need an easy way to just tell Call whether or not RtpTransport is in use, you could add some flag to CallConfig.

We don't know that until we set an offer.

At this point I am basically avoiding doing large refactorings because
1) unit tests are not very comprehensive
2) there are hidden assumptions in downstream clients (like caching of the call object to increase performance) that I might not be aware of.

I can however see if we can move the 'AddTargetRateObserver' callback to the first 'OnTransportCreated' callback if that's something we strongly prefer.

Patch-set: 2
Reviewer: Gerrit User 7704 <7704@58829da1-049c-39fb-b951-ebdcd0984767>