Update patch set 7
Patch Set 7:

(6 comments)

Looks quite reasonable, although I'm still a bit confused as to when the MediaTransportRange method is invoked.

It would be cleaner to also have some method called when the decision to use RTP transport is made, insteadof calling RegisterRateObserver when each stream is created and checking a flag to see if it's the first stream. Which looks like a kludge. I'm not familiar with the OnTransportChange logic though, so I can't say if and where such a callback would fit.

Patch-set: 7
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>