Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (6 comments)
> 
> Looks quite reasonable, although I'm still a bit confused as to when the MediaTransportRange method is invoked.
  // Gets called when media transport is created or removed.


> 
> It would be cleaner to also have some method called when the decision to use RTP transport is made, insteadof calling RegisterRateObserver when each stream is created and checking a flag to see if it's the first stream. Which looks like a kludge. I'm not familiar with the OnTransportChange logic though, so I can't say if and where such a callback would fit.

There is no place where we know that RTP is going to be used, only a place where we know it won't (that is, if MediaTransport is registered, we know that RTP is going to be used).

We could add a method 'TransportSelectedCallback' and propagate RTP/MediaTransport enum, but given that 1) RTP codepath is going away when we move to RTP media transport 2) there are multiple callbacks in jseptransportcontroller (not just one), I didn't want to do it.

Patch-set: 7
Reviewer: Gerrit User 7704 <7704@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0