Update patch set 9
Patch Set 9:

It would be good to make sure this is run on some bot so that we'll catch regressions, but apart from that, is there a need to have this support other operating systems than WinUWP, which is the main goal?

Personally, I would like to continue to look into leveraging platform capabilities wherever possible (e.g. see the GCD implementation). I know we've discussed that option briefly here and decided that there's not an obvious fit we're aware of right now, but there could be in the future. I'd like to leave the door open to that and would also like to avoid someone introducing a dependency on this implementation for some unforeseen convenience, on non-WinUWP platforms, just because it's possible.

Patch-set: 9
Reviewer: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>