Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> It would be good to make sure this is run on some bot so that we'll catch regressions, but apart from that, is there a need to have this support other operating systems than WinUWP, which is the main goal?

I'm thinking both of bots, and local testing. If there's some breakage related to this code, more people could debug it if it didn't require a windows development machine. A  single gn flag should be enough for both use cases.

I don't expect us to use the implementation in production on anything but WinUWP in the near future,but it would be interesting to compare performance and binary size of this task queue to the libevent-based one. I'm not aware of any linux-specific threadpool things that are likely to be more efficient than using std::thread facilities.

One thing that we have talked about for some time is the need to have cheaper task queues on linux and android, using a thread pool instead of one dedicated thread per tq. I don't know what our best options are, but if we decide to add our own thread pool instead of libevent, it might make sense to use the same also on windows, possible extending this code.

Patch-set: 9
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>