Update patch set 8
Patch Set 8:

(11 comments)

Hi,
Thanks for the CL! I think it makes a lot of sense to do this move.
I have some comments but a general one is this.

This CL moves the pre-amplifier gain into AGC2 but without adding that much functionality yet. This move causes problems with the settings (as you pointed out).

Would it be possible to instead maintain both functionalities (the old and the new) active in parallel until the AGC2 support for this properly has landed?
That would solve the difficulties with the settings.

Patch-set: 8
Reviewer: Gerrit User 5125 <5125@58829da1-049c-39fb-b951-ebdcd0984767>