AEC3: Optimizing the Update method of the FilterAnalyzer class.
In this CL the analysis of the impulse response that is done in the FilterAnalyzed class is changed in order to reduce its complexity. Instead of analyzing the whole impulse response in each Update call a smaller region is analyzed. That region is changed at each Update call which implies that several calls are needed in order to analyze the complete impulse response.

Bug: webrtc:10032
Change-Id: Ic58be34ba18485311c63e0fed9b6e892f9cb864c