Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Would another option be to have the internal class simply implement DtlsTransportInterface?

That would be an option - but the stuff that needs to be exposed externally (chiefly an event listener) and the stuff that needs to be done internally seem to be very disjunct fuction sets, so I thought the minimally invasive way was this wrapping.

We can merge them later if it turns out I was wrong about that.

Patch-set: 1
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>