Update patch set 8
Patch Set 8:

(14 comments)

Awesome! Thanks for doing this CL!!!! 
I have some comments on the code.

Also, regarding the CL description please
-Make the title shorter to fit a single line.
-Add a description that describes a bit more in detail what is accomplished by the CL. In particular, I think there should be a comment that states that this the filter is analyzed in a cyclic manner.


Also, since a kill-switch is added for this, have you checked for bit-exactness with the former code when the kill-switch is on?

Patch-set: 8
Reviewer: Gerrit User 5125 <5125@58829da1-049c-39fb-b951-ebdcd0984767>