Update patch set 9
Patch Set 9:

(1 comment)

> Patch Set 8:
> 
> (6 comments)
> 
> I guess we could as a separate step refactor all sample counters to be int64_t instead.

That would be ideal. However, my nagging is motivated by a sense that many such future refactorings fail to occur, so it's often better to at least try to ensure that new code does it the right way.

But if you're willing to pinky promise to actually do such a refactoring, I could be convinced to take it easy on you in this review. :-)

Patch-set: 9
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>