Don't clear self.videoFrame when setting up OpenGL in the EAGL video view.
It makes sense to clean up self.videoFrame in -teardownGL, but if
we happen to have a frame available in -setupGL then it's OK to
keep using that frame. Clearing the frame here frequently causes
the screen view to go black for a moment when the app returns from
the background.

Bug: 10059
Change-Id: Ic62f872a0a582c807cee1e30ea9bb32f31ada341