Update patch set 1
Patch Set 1: Code-Review+1

I think this is right. The important thing is, *if* we for some reason need some magic value internal to the implementation meaning "not set" or the like, that must not be mixed up with values that can appear on the wire.

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1