AEC3: HMM implementation of dominant nearend detection
Adds a new implementation of the dominant nearend detection. The new
classifier detects nearend faster and more reliably. Better nearend
detection enables tuning the suppressor for better transparency during
double-talk.

The new classifier uses a hidden markov model to estimate the probability
of being in "nearend" vs "normal" state. Observable output is
echo-to-nearend ratio and signal-to-noise ratio.

The old implementation is available and can be activated in the
configuration struct or by a kill-switch.

Change-Id: Ia7d3f1b61c7807936f36985a3fbe100055f7bf68