Update patch set 2
Patch Set 2:

(1 comment)

Not yet working, but maybe you would like to have an initial look? The pattern is going to be common: A test uses a special encoder/decoder, injects it into component under test wrapped in a factory. And it then needs to examine internals of the special encoder/decoder.

For video we're now doing that using the VideoDecoderProxyFactory, which is given a single decoder instancs, and the factory's create method hands out proxies (as unique_ptr:s) all delagating to that same decoder. This cl tries to do the same for one of the NetEq tests.

Patch-set: 2
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5100 <5100@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false