Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Julien, please help to take a look. And any better idea since this will impact the performance a bit?

Hi,LGTM. I cannot think of a better solution. Also the perf impact will not happen if zoom 1x1, i.e. zoom enabled but zoomed all the way out as mentioned in UAZoomEnabled doc.

Worth to mention that we can avoid the frame copy over (prev frame to current frame) when invalidating the whole region , see one of the TODO in same file, which refers to obsolete bug or comment not accurate. Search queue and memcpy. So we should be able to get back the 5-10 fps lost here. But something to tackle separately.

Patch-set: 1
Reviewer: Gerrit User 6236 <6236@58829da1-049c-39fb-b951-ebdcd0984767>