Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I suggest reverting the previous commit if the behavior is currently broken. Then reconsider if it's reasonable to defer the decision on whether to use rtptransportcontrollersend or media transport if it needs to report bitrates early.

The usual way to handle downstream breakage would be to (1) revert original cl, (2) fix downstream clients (in this case, chromium?), reland cl. Do you know if it's hard or easy to fix downstream? Is it just the way things are setup in tests, or do we expect breakage also in production use?

I'm also a bit curious what would happen if we unconditionally subscribe to the RTP BWE estimator (without the wrapper class), and in addition subscribe to MediaTransport. Will the RTP code generate any callbacks if it isn't used?

Patch-set: 4
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>