Update patch set 5
Patch Set 5:

(1 comment)

Which seems to mean that NetEq doesn't call the Decode methods directly. It would be good to document in audio_decoder.h that ParsePayload is the advertised external entrypoint, not the Decode* methods? That's a bit surprising to a first-time reader of the code.

Seems to work fine now.

Patch-set: 5
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>