Update patch set 1
Patch Set 1: Commit-Queue+2

> Patch Set 1: Code-Review+1
> 
> > Patch Set 1:
> > 
> > Is this better?
> 
> Yes, it's a bit clearer. Maybe mention that the main use case is when the other endpoint isn't supporting webrtc's more general bwe?

That's an explanation for when it would be *useful* to make use of the codec-internal (and codec-specified) BWE. The business of the docs here is to explain how to set up encoder-decoder communications; giving codec BWE as an example of when that would be useful is already a side note, and explaining when codec BWE would be useful would be a side note in the side note.

If you feel strongly about it I can of course add it, but I don't feel it would be a good fit for terse reference documentation like this.

Patch-set: 1
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=+2