Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is this leading into a broader change across the entire codebase (accompanied with a presubmit)?

I would like to say "yes" but I don't know yet. :)
I am trying to untangle a problem (and it turns out this is not enough so I need to re-work this CL) that I want to fix before the end of the week.

My long term vision here is to use clang-tidy for these kind of things (instead of regex in PRESUBMIT.py), yvesg@ is exploring the possibility of bringing more static analysis into WebRTC. If that will happen then I think a broader cleanup will be the next step.

Patch-set: 1
Reviewer: Gerrit User 5087 <5087@58829da1-049c-39fb-b951-ebdcd0984767>