Desktop capturer: Add OnDisplayChanged callback
This adds support for a new DesktopCapturer::Callback method
OnDisplayChanged that is sent at the start of a desktop capture
session and whenever the display geometry changes.

This cl adds the basic structure to call this api at the start
of the capture session. Currently Windows only.

A follow-up cl will add support to call this whenever the display
geometry changes.

Bug: webrtc/10122, chromium/915411
Change-Id: Ie7283be5992454180daab1a60f58a3b2efdfed56