Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Sebastian: do we want to continue trying to figure out what's going on, or for the sake of the clients, should we check it in in the meantime?

I don't think this solution is appropriate as motivated before. If we have identified a regression in WebRTC behavior due to the previous CL we should try to find exactly when/why it happens. If we do that, I suspect that we will be able to find a solution that does not require registering as observer on construction.

If the problem is such that we fundamentally need to register as observer from the start, and we want to preserve the behavior, we will have to change the design for integrating MediaTransport so it can be set at construction time. We might then want to revert the current integration pending that solution, but hopefully there's a better solution per above.

Patch-set: 4
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>