Update patch set 5
Patch Set 5:

(6 comments)

Talked offline, and this sounds like the best plan moving forward with a fix that isn't error prone for getting into M72. Overall looks good for the given plan. I think it's worth adding a TODO somewhere with a reference to the bug, so we could potentially follow up on this later with a more robust refactor/fix.

Patch-set: 5
Reviewer: Gerrit User 5800 <5800@58829da1-049c-39fb-b951-ebdcd0984767>