Update patch set 9
Patch Set 9:

(3 comments)

Looks OK to me in general.

I get a bit confused when we have a variable and a bool that says whether it's there or not - I prefer to use rtc::optional for the cases where we need to do this.
If we want to have has_mid say something about the remote SDP only, and still have mid() give back the locally assigned name (always present), we might want to resurrect "mid" and "name" as distinct concepts.

I couldn't tell if you treat the case of a new (updated) offer specially; I think we'd want to reject if we had assigned a mid because it was missing and an updated offer came along with a conflicting mid value - might be worth writing a test for.

Patch-set: 9