Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)
> 
> I don't think this is the right place to do it. Copying pixels around belongs close to he capturer, and the capturer should know what type of frame buffer it needs, and e.g., keep a I420Buffer that it can draw into using the MutableData* accessors.
> 
> The more abstract VideoFrameBuffer interface is intended to be immutable.

This is after the capturer, so we can pass a partially changed frame to webrtc. Without changing the VideoFrameSink interface, VideoFrame will have an optional struct telling that it's actually a part of the screen. This will be set by capturer. And in WebRTC we will paste that onto last image using this new routines.

The VideoFrameBuffer isn't immutable as we already do some stuff there, like rotation and scaling.

Patch-set: 1
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0