Update patch set 5
Patch Set 5:

(5 comments)

Great initiative!

I see that the interface is similar to the Scenario class, would it make sense to share the network emulation implementation?

Do we really need to add so much to API? Maybe we could start working in /test and move to aPI when we think it's ready.

Is the manager really peerconnection specific? Maybe we could move the files added in /api/test and /test/pc/e2e to something like /test/network?

Patch-set: 5
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>