Update patch set 5
Patch Set 5:

(5 comments)

About sharing with scenario - here we need a bit more information, then there. Because it supposed to be used on system socket layer we need ip addresses to correctly route packets, so it will require introduction of unnecessary parts into scenario code..

About adding to API -  we need that downstream projects will be able to construct required simulation in there tests, so construction API should be public. Also they need to be able to implement there own nodes, so node definition also should be public.

> Is the manager really peerconnection specific? Maybe we could move the files added in /api/test and /test/pc/e2e to something like /test/network?

Maybe we can. The main point, that it will be injected into network thread, and to access it easily you need whole transport stack. Maybe in some cases you can construct it in some way to be able to inject network manager. Or maybe it will be possible to reuse parts of it in some cases. I can't tell right now. But probably we can put it into /test/network..

Should I?

Patch-set: 5
Label: Commit-Queue=0