Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> This seems like it would affect the actual behavior of WebRTC? Shouldn't it be under field trial to ensure that we don't have regressions?

The actual bitrate this subtracts is on the order of 400 to 3000bps depending on packet rate, in my tests. So the effect will be extremely subtle unless the pacing factor is at 1.0 (and even then).

Nevertheless having a kill-switch is probably a good idea. Adding one here:
https://webrtc-review.googlesource.com/c/src/+/116661

Patch-set: 8