Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think you relanded a broken commit without any changes..?
> Also bypassing all tests (https://bugs.chromium.org/p/chromium/issues/detail?id=895971)

Oh, didn't realize it'd skip the tests.

But from the thread on it, the brokeness didn't seem to be related to the commit as rolling it back did not help? As the CL doesn't change any behavior (it only adds code) it seems odd that it would break tests?

Patch-set: 2