Fixing possible bug when Flex and RTX used together.
When Flex and RTX are both specified, Flex will not be used because
RTX will introduce a new SSRC making the total SSRC count > 1.
Flex can only protect a single stream, so if the total SSRC count
is > 1, it is not used.
The fix is simple, to check the number of "primary SSRCs" before
redundancy streams are added when determining if Flex should be used.

Bug: None
Change-Id: I98df1b807d306bdcce1a76dfb163aa14e60d0052