Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not really familiar with this code. I agree the use of the static variable can have surprising and undesired effects. A few comments:
> 
> One way of fixing the problem is to add an argument to the RTC_HISTOGRAM_COMMON_BLOCK area that provides either a name for the static variable, or a pointer to the location where it should be cached. Or take out the caching to separate macro and pass the histogram pointer (cached or not) to RTC_HISTOGRAM_COMMON_BLOCK. The point is to give the caller of the macro control of the caching, and if/how the cached data is shared.
> 
> Second, if a static variable is used, explicit atomic ops are no longer needed, we can rely on static initializers to be thread safe from C++11.

Thanks for the tips! I agree a thorough fix need a rewrite of the macro, particularly how the caching is done. The current fix is to restore the intended behavior, and hopefully recover the metrics in the upcoming Chromium milestones.

Do you think if we can defer the refactoring of the macro to a follow-up CL?

Patch-set: 1
Reviewer: Gerrit User 5887 <5887@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0