Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > This change is ready for review.
> 
> Just to make sure I understand the background before I stamp on it: removing these for-test-only methods assumes our test infra now supports IPv6. Is this true now? Looks like all trybots are green, but just want to double check.

I think we can now rely on an infra that supports IPv6 but I am not 100% sure about that. Green trybots are a good first step. If you agree we can land this one and keep an eye on it. If something goes wrong and we hit the problem I can revert.

Patch-set: 3
Reviewer: Gerrit User 5087 <5087@58829da1-049c-39fb-b951-ebdcd0984767>