Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5:
> 
> (1 comment)
> 
> > Patch Set 5:
> > 
> > (4 comments)
> > 
> > Agree with the brittleness issue. However, there's a dependency graph problem.
> > 
> > Suggest landing this solution and iterating.
> 
> Isn't is possible to break the dependency cycle by refactoring p2p:rtc_p2p?

Anything is possible - the question is whether to require doing it before exposing
webrtc::IceTransport to Chromium, or whether to keep the current hierarchy of containment
and ownership for now and redesign it later.

> Maybe we can split it in more targets or maybe we need to rethink api:libjingle_peerconnection_api, but in both cases, landing this CL as is will introduce two problems:
> 1. The aforementioned issue with declarations and definitions in 2 different build targets.

That's already fixed.

> 2. It is not clear if we want a big API target or if we want big internal targets, having both will generate circular deps, so we need to decide what to do there.

There are no circular dependencies in this CL. It's the possible refactoring of p2p that could introduce a circular dependency between webrtc/api and p2p.

Patch-set: 5
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>