Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 5:
> > 
> > (1 comment)
> > 
> > > Patch Set 5:
> > > 
> > > (1 comment)
> > > 
> > > > Patch Set 5:
> > > > 
> > > > (4 comments)
> > > > 
> > > > Agree with the brittleness issue. However, there's a dependency graph problem.
> > > > 
> > > > Suggest landing this solution and iterating.
> > > 
> > > Isn't is possible to break the dependency cycle by refactoring p2p:rtc_p2p?
> > 
> > Anything is possible - the question is whether to require doing it before exposing
> > webrtc::IceTransport to Chromium, or whether to keep the current hierarchy of containment
> > and ownership for now and redesign it later.
> 
> I am ok with redesign later only if we have a clear plan. Sorry for being so nitpicking but api/ is supposed to be our last big organizational refactoring and if we don't have a clear plan to reduce technical debt when we buy it then we will end up in an unsatisfying situation (consider that we are already trying to fix some well known issues, like other violations of https://webrtc.googlesource.com/src/+/HEAD/style-guide.md#h-cc-pairs and usage of // nogncheck).
> 
> > > Maybe we can split it in more targets or maybe we need to rethink api:libjingle_peerconnection_api, but in both cases, landing this CL as is will introduce two problems:
> > > 1. The aforementioned issue with declarations and definitions in 2 different build targets.
> > 
> > That's already fixed.
> 
> I am not sure it is fixed, isn't webrtc::CreateIceTransport declared by api:libjingle_peerconnection_api and defined by api:ice_transport_api?

That's true. 
> 
> > > 2. It is not clear if we want a big API target or if we want big internal targets, having both will generate circular deps, so we need to decide what to do there.
> > 
> > There are no circular dependencies in this CL. It's the possible refactoring of p2p that could introduce a circular dependency between webrtc/api and p2p.
> 
> Yes, this CL is ok from that point of view. I was thinking that ice_transport_interface.cc is in its own build target because it depends on p2p:rtc_p2p (which depends on api:libjingle_peerconnection_api), so the decision to violate https://webrtc.googlesource.com/src/+/HEAD/style-guide.md#h-cc-pairs is caused by the potential circular dependency, is this assumption correct?

I don't think it's a potential circular dependency, it's fear of touching big central targets.

I was afraid of letting libjingle_peerconnection_api depend on p2p, because I could see from the dependency graph that it currently didn't. But anyone using Peerconnection will end up having to depend on something that includes p2p, so I figured that I don't understand what the reasoning around the libjingle_peerconnection_api target is, so I don't want to touch it more than I have to.

api/peer_connection_interface.h refers to api/dtls_transport_interface.h which refers to api/ice_transport_interface.h - so I thought they all had to be in the same target, or in a dependency of that target. True or false?

(I assume moving peer_connection_interface.h out of libjingle_peerconnection_api is not on the books...)

I'll take any proposed change to the BUILD.gn and DEPS files as long as a) it compiles, and b) it passes the bots. If in addition I understand why it's the best solution, that's an added bonus.

Patch-set: 6
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>